Changed vehicle logic  and reformatted the project.
Main screen (vehicle list) shows only Make and Model. Clicking on a vehicle displays more information about the vehicle in a new screen. Also in the new screen there is an update button and a delete button with prompt. Insert/Update is the same screen but with other parameters. Also changed the way vehicletypes are loaded for now.